backend/requirements.txt
Django==4.2.29
boto3==1.42.96
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-colorfield==0.14.0
django-timezone-field
dj-database-url==3.0.1
psycopg[binary]==3.2.13
gunicorn==23.0.0
whitenoise==6.11.0
pillow==11.3.0
pypdf==5.9.0
python-dotenv
